pyproject.toml
[project]
name = "pbl-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gdown>=5.2.0",
    "huggingface-hub>=0.36.0",
    "hydra-core>=1.3.2",
    "librosa>=0.11.0",
    "nemo-toolkit[tts]==2.5.0",
    "numpy>=1.26.4",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
]
